Refactor: Move source map logic to context.ts
